# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Mangue Space Coworking - Módulo de Endereço Fiscail",
        "description": "<strong>A API(Application Programing Interface) de Endereço Fiscal</strong> é um <strong>módulo do Sistema de Gerenciamento do Coworking Mangue Space</strong>, esse módulo será responsável por: <ul><li>Cadastramento e gerenciamento das empresas contratantes do serviço da Softex;</li><li>Organizar em um sistema a listagem das empresas com cadastro ativo aptos para recebimento de encomendas e correspondências;</li><li>Notificar os Representantes das empresas cadastradas após o recebimento de encomendas e correpondências.</li></ul>",
        "termsOfService": "http://softex.com.br",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API de teste"
        }
    ],
    "paths": {
...
        "/cliente/ativados": {
            "get": {
                "tags": ["Cliente"],
                "summary": "Listagem de Clientes Ativos",
                "description": "Deverá ser possível que um <strong>usuário</strong> possa listar todos os Clientes que estão cadastrados e ativos na base de dados do sistema.",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Não Autorizado!"
                    }
                }
            }
        },
        "/cliente/inativados": {
            "get": {
                "tags": ["Cliente"],
                "summary": "Listagem de Clientes Inativos",
                "description": "Deverá ser possível que um <strong>usuário administrador</strong> possa listar todos os Clientes que estão cadastrados e inativos, na base de dados do sistema.",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Não Autorizado!"
                    }
                }
            }
        },
...
        "/cliente/id/{id}": {
            "get": {
                "tags": ["Cliente"],
                "summary": "Busca de Cliente por ID",
                "description": "Deverá ser possível que um <strong>usuário</strong> possa buscar um Cliente pelo número do ID na base de dados do sistema.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Cliente para busca",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cliente não encontrado!"
                    },
                    "401": {
                        "description": "Não Autorizado!"
                    }
                }
            }
        },
...
        "/recebimentoencomenda/": {
            "post": {
                "tags": ["Recebimento de Encomenda"],
                "summary": "Cadastramento de Recebimento de Encomendas",
                "description": "O recebimento de encomendas acontece na recepção do ITBC. Um <strong>usuário</strong> da recepção deverá identificar se existe um Cliente ativo para poder receber a encomenda recém chegada. Após a verificação e constatação de cadastro ativo, o usuário terminará o recebimento e o representante cadastrado no Cliente, receberá uma notificação via Whatsapp e E-mail.",
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "example": {
                                "obsEncomenda": "Encomenda um pouco avariada nas extremidades",
                                "obsRecebEncomenda": "Entregue pelos Correios",
                                "numEndFiscal": 44,
                                "idRecepcao": 2
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recebimento de Encomenda cadastrado!"
                    },
                    "400": {
                        "description": "Requisição Inválida!"
                    },
                    "401": {
                        "description": "Não Autorizado!"
                    }
                }
            },
            "get": {
                "tags": ["Recebimento de Encomenda"],
                "summary": "Listagem dos Recebimentos de Encomendas",
                "description": "Deverá ser possível que um <strong>usuário</strong> possa listar todos os Recebimentos de Encomendas que estão cadastrados, na base de dados do sistema.",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Não Autorizado!"
                    }
                }
            }
        },
        "/recebimentoencomenda/id/{id}": {
            "get": {
                "tags": ["Recebimento de Encomenda"],
                "summary": "Busca de Recebimento de Encomenda por ID",
                "description": "Deverá ser possível que um <strong>usuário</strong> possa buscar um Recebimento de Encomenda pelo número do ID na base de dados do sistema.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Recebimento de Encomenda para busca",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecebimentoEncomenda"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Recebimendo de Encomenda não encontrado!"
                    },
                    "401": {
                        "description": "Não Autorizado!"
                    }
                }
            }
        },
        "/retiradaencomenda/": {
            "post": {
                "tags": ["Retirada Encomenda"],
                "summary": "Cadastramento de Retirada de Encomendas",
                "description": "A retirada de encomendas acontece na recepção do ITBC. Um <strong>representante</strong> do Cliente PJ deverá identificar se existe um Cliente ativo para poder retirar a encomenda recém chegada. Após a verificação e constatação de cadastro ativo, o representante fará a retirada da encomenda.",
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "example": {
                                "idRetirEncomenda": 14,
                                "dataHoraRetirEncomenda": "2024-01-09T22:34:54.000Z",
                                "obsRetirEncomenda": "Entregue pelos Correios",
                                "idEncomenda": 6,
                                "idRepresent": 2
                              
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retirada de Encomenda cadastrado!"
                    },
                    "400": {
                        "description": "Requisição Inválida!"
                    },
                    "401": {
                        "description": "Não Autorizado!"
                    }
                }
            },
            "get": {
                "tags": ["Retirada Encomenda"],
                "summary": "Listagem dos Retiradas de Encomendas",
                "description": "Deverá ser possível que um <strong>representante</strong> possa listar todos as Retiradas de Encomendas que estão cadastrados, na base de dados do sistema.",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Não Autorizado!"
                    }
                }
            }
        },
        "/retiradaencomenda/id/{id}": {
            "get": {
                "tags": ["Retirada Encomenda"],
                "summary": "Busca de Retirada de Encomenda por ID",
                "description": "Deverá ser possível que um <strong>representante</strong> possa buscar um Retirada de Encomenda pelo número do ID na base de dados do sistema.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Retirada de Encomenda para busca",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RetiradaEncomenda"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Retirada de Encomenda não encontrado!"
                    },
                    "401": {
                        "description": "Não Autorizado!"
                    }
                }
            }
        }   
    },
    "components": {
        "schemas": {
...
            "Cliente": {
                "type": "object",
                "properties": {                    
                    "statusCliente": {
                        "type": "number"
                    },
                    "createdAtCliente": {
                        "type": "datetime"
                    },
                    "updatedAtCliente": {
                        "type": "datetime"
                    },
                    "idCliente": {
                        "type": "number"
                    },
                    "nomeCliente": {
                        "type": "string"
                    },
                    "telefoneCliente": {
                        "type": "string"
                    },
                    "emailCliente": {
                        "type": "string"
                    },
                    "qtdPontosCliente": {
                        "type": "number"
                    },
                    "prazoCliente": {
                        "type": "date"
                    },
                    "enderecoIdEndereco": {
                        "type": "number"
                    },
                    "adminIdAdmin": {
                        "type": "number"
                    },
                    "enderecoFiscalNumEndFiscal": {
                        "type": "number"
                    }
                }
            },
...
            "RecebimentoEncomenda": {
                "type": "object",
                "properties": {                    
                    "dataHoraRecebEncomenda": {
                        "type": "datetime"
                    },
                    "idRecebEncomenda": {
                        "type": "number"
                    },
                    "obsRecebEncomenda": {
                        "type": "string"
                    },
                    "encomendaIdEncomenda": {
                        "type": "number"
                    },
                    "enderecoFiscalNumEndFiscal": {
                        "type": "number"
                    },
                    "idRecepcao": {
                        "type": "number"
                    }
                }
            },
            "RetiradaEncomenda": {
                "type": "object",
                "properties": {                    
                    "dataHoraRetirEncomenda": {
                        "type": "datetime"
                    },
                    "idRetirEncomenda": {
                        "type": "number"
                    },
                    "obsRetirEncomenda": {
                        "type": "string"
                    },
                    "idEncomenda": {
                        "type": "number"
                    },
                    "idRepresent": {
                        "type": "number"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}